+++ srcs/time.c
@@ -6,15 +6,13 @@
 /*   By: lbouchon <lbouchon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/22 11:18:29 by lbouchon          #+#    #+#             */
-/*   Updated: 2023/03/13 13:10:10 by lbouchon         ###   ########.fr       */
+/*   Updated: 2023/03/13 17:43:13 by lbouchon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
-#include <unistd.h>
-#include <sys/time.h>
+#include "../include/philo.h"
 
-long int	get_time(void)
+int	get_time(void)
 {
 	struct timeval	tp;
 
@@ -22,9 +20,9 @@ long int	get_time(void)
 	return ((tp.tv_sec * 1000) + (tp.tv_usec / 1000));
 }
 
-void	ft_usleep(long int time_in_ms)
+void	ft_usleep(int time_in_ms)
 {
-	long int	start_time;
+	int	start_time;
 	
 	start_time = get_time();
 	while ((get_time() - start_time) < time_in_ms)
